feat(comment) : 게시물 상세 조회 댓글 유저 반환

+++ src/main/kotlin/com/example/potpot/domain/apply/domain/ApplyRepository.kt
@@ -9,7 +9,7 @@ interface ApplyRepository : JpaRepository<Apply, Long> {
 
     fun findByFeed(feed: Feed): List<Apply>
 
-    fun findByFeedIdAndUserId(feedId: Long, userId: Int): Apply?
+    fun findByFeedIdAndUserId(feedId: Long, userId: Long): Apply?
 
     fun existsByUserAndFeed(user: User, feed: Feed): Boolean
 }
